#14 AppSyncリクエスト関数をutilsにまとめた
